+++ lib/routes/withdraw_funds/sweep_all_coins_confirmation.dart
