+++ lib/user_app.dart
@@ -14,8 +14,8 @@ import 'routes/security_pin/lock_screen.dart';
 import 'routes/add_funds/deposit_to_btc_address_page.dart';
 import 'routes/add_funds/moonpay_webview.dart';
 import 'routes/get_refund/get_refund_page.dart';
+import 'routes/withdraw_funds/unexpected_funds.dart';
 import 'routes/withdraw_funds/reverse_swap_page.dart';
-import 'routes/withdraw_funds/send_coins_dialog.dart';
 import 'package:breez/widgets/static_loader.dart';
 import 'package:flutter/cupertino.dart';
 import 'routes/connect_to_pay/connect_to_pay_page.dart';
